Borders are now enforced.
